requirements.txt
discord.py>=2.5,<3.0
python-dotenv>=1.1,<2.0
requests>=2.32,<3.0